+++ tests/class.global.test.mjs
@@ -9,4 +9,5 @@ test('class.global', () => {
     expect(method_exists).not.toBeUndefined();
     expect(property_exists).not.toBeUndefined();
     expect(get_object_vars).not.toBeUndefined();
+    expect(get_class_vars).not.toBeUndefined();
 });